MAJOR MEMORY REDUCTIONS
Made major changes to all variables & constants to reduce overal global memory usage by over 18%. These changes will also impact local variables but no "EASY" way to measure that impact. Still more that could be done but this was a huge overhaul to make everything take less space. (Ex. Morse went from using 1265 bytes down to only 589)

+++ IRLKTANE/batteries.h
@@ -5,4 +5,4 @@
   KNOWN ISSUES:
   Will eventually become a random number of batteries and actualy display them. For now it is fixed values
 */
-int nrBatteries = 0; // number of batteries // will eventually point to a .h file that controls which batteries are shown
+byte nrBatteries = 0; // number of batteries // will eventually point to a .h file that controls which batteries are shown
